refactor StructsToJson handling
