+++ spark/src/main/scala/org/apache/comet/serde/json.scala
@@ -1,0 +1,100 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.comet.serde
+
+import org.apache.spark.sql.catalyst.expressions.{Attribute, Expression, StructsToJson}
+import org.apache.spark.sql.types._
+
+import org.apache.comet.CometSparkSessionExtensions.withInfo
+import org.apache.comet.serde.ExprOuterClass
+import org.apache.comet.serde.QueryPlanSerde.exprToProtoInternal
+
+object CometStructsToJson extends CometExpressionSerde {
+
+  override def convert(
+      expr: Expression,
+      inputs: Seq[Attribute],
+      binding: Boolean): Option[ExprOuterClass.Expr] = {
+    expr match {
+      case StructsToJson(options, child, timezoneId) =>
+        if (options.nonEmpty) {
+          withInfo(expr, "StructsToJson with options is not supported")
+          None
+        } else {
+
+          def isSupportedType(dt: DataType): Boolean = {
+            dt match {
+              case StructType(fields) =>
+                fields.forall(f => isSupportedType(f.dataType))
+              case DataTypes.BooleanType | DataTypes.ByteType | DataTypes.ShortType |
+                  DataTypes.IntegerType | DataTypes.LongType | DataTypes.FloatType |
+                  DataTypes.DoubleType | DataTypes.StringType =>
+                true
+              case DataTypes.DateType | DataTypes.TimestampType =>
+                // TODO implement these types with tests for formatting options and timezone
+                false
+              case _: MapType | _: ArrayType =>
+                // Spark supports map and array in StructsToJson but this is not yet
+                // implemented in Comet
+                false
+              case _ => false
+            }
+          }
+
+          val isSupported = child.dataType match {
+            case s: StructType =>
+              s.fields.forall(f => isSupportedType(f.dataType))
+            case _: MapType | _: ArrayType =>
+              // Spark supports map and array in StructsToJson but this is not yet
+              // implemented in Comet
+              false
+            case _ =>
+              false
+          }
+
+          if (isSupported) {
+            exprToProtoInternal(child, inputs, binding) match {
+              case Some(p) =>
+                val toJson = ExprOuterClass.ToJson
+                  .newBuilder()
+                  .setChild(p)
+                  .setTimezone(timezoneId.getOrElse("UTC"))
+                  .setIgnoreNullFields(true)
+                  .build()
+                Some(
+                  ExprOuterClass.Expr
+                    .newBuilder()
+                    .setToJson(toJson)
+                    .build())
+              case _ =>
+                withInfo(expr, child)
+                None
+            }
+          } else {
+            withInfo(expr, "Unsupported data type", child)
+            None
+          }
+        }
+      case _ =>
+        withInfo(expr, s"Unexpected expression type for CometStructsToJson: ${expr.getClass}")
+        None
+    }
+  }
+}
